requirement.txt
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
packaging==23.1
pkgutil-resolve-name==1.3.10
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.29.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
